+++ LastGame/src/main/java/com/IGallinari/LastGame/payload/response/DetailsPlayerIndependByGame/ViewShotsDetailsPlayerIndependentByGame.java
@@ -8,11 +8,11 @@ import lombok.Data;
 public class ViewShotsDetailsPlayerIndependentByGame {
     private Integer fieldShotsScored;
     private Integer fieldShotsAttempted;
-    private Integer fieldShotsPercentage;
+    private Float fieldShotsPercentage;
     private Integer freeThrowsScored;
     private Integer freeThrowsAttempted;
-    private Integer freeThrowsPercentage;
+    private Float freeThrowsPercentage;
     private Integer threePointersShotsScored;
     private Integer threePointersShotsAttempted;
-    private Integer threePointersShotsPercentage;
+    private Float threePointersShotsPercentage;
 }
